podiobooks/requirements_staging.txt
-r requirements.txt
django-redis-cache
django-redisboard
django-redis-sessions
psycopg2
hiredis
redis